Update HP datatype

+++ Enemies/Bird.cs
@@ -4,8 +4,8 @@ public class Bird : IEnemy
 {
     public string Name { get; } = "Birdy";
     public string ImagePath { get; } = "img/bird.png";
-    public int Health { get; set; } = 60;
-    public int MaxHealth { get; } = 60;
+    public float Health { get; set; } = 60;
+    public float MaxHealth { get; } = 60;
     public int Level { get; } = 1;
     public List<string> Noise { get; } = new List<string>()
     {
