+++ Enemies/Tree.cs
@@ -2,8 +2,8 @@ namespace Game.Enemies;
 class Tree : IEnemy {
 public string Name { get; } = "Tree";
 public string ImagePath { get; } = "img/tree.png";
-public int Health { get; set; } = 100;
-public int MaxHealth { get; } = 100;
+public float Health { get; set; } = 100;
+public float MaxHealth { get; } = 100;
 public int Level { get; } = 1;
 public List<string> Noise { get; } = new List<string>() {
 "tree",
